+++ CHANGELOG.md
@@ -3,8 +3,9 @@
 ### master
 
 - Fixed: Firing sequence of `common_toolkit_loaded` hook
-- Added: Ability to modify or disable WordPress heartbeat
 - Added: `ctk_environment` filter
+- Added: Ability to disable WordPress core, plugin and/or theme updates
+- Added: Ability to modify or disable WordPress heartbeat
 - Added: Ability to disable WordPress search
 - Added: Ability to define custom environment constant